feat: Added repo for comments and endpoints for manipulation of comments refactor: changed value in repos to long

+++ src/main/java/se/experis/noticeboard/Repositories/NoticeRepository.java
@@ -3,5 +3,5 @@ package se.experis.noticeboard.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import se.experis.noticeboard.models.Notice;
 
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 }
